refactor: rename file of PRVI iterator according to naming standards
