+++ src/queens/iterators/__init__.py
@@ -40,7 +40,7 @@ if TYPE_CHECKING:
     from queens.iterators.points import Points
     from queens.iterators.polynomial_chaos import PolynomialChaos
     from queens.iterators.reinforcement_learning import ReinforcementLearning
-    from queens.iterators.reparameteriztion_based_variational import RPVI
+    from queens.iterators.rpvi import RPVI
     from queens.iterators.sequential_monte_carlo import SequentialMonteCarlo
     from queens.iterators.sequential_monte_carlo_chopin import SequentialMonteCarloChopin
     from queens.iterators.sobol_index import SobolIndex
